refactor(peer): make domain own room contracts
